+++ src/agenticcybersense/mcp/functions/rag.py
@@ -1,0 +1,1 @@
+"""Generate a response using Retrieval-Augmented Generation (RAG)"""
